jmap: refactor email body UTF-8 decoder and add confidence parameter
