+++ imap/jmap_util.h
@@ -112,4 +112,25 @@ extern json_t *jmap_server_error(int r);
 extern char *jmap_encode_base64_nopad(const char *data, size_t len);
 extern char *jmap_decode_base64_nopad(const char *b64, size_t b64len);
 
+/* Decode the text in data of datalen bytes to UTF-8 to a C-string.
+ *
+ * Attempt to detect the right character encoding if conversion to
+ * UTF-8 yields any invalid or replacement characters.
+ *
+ * Parameters:
+ * - charset indicates the presumed character encoding.
+ * - encoding must be one of the encodings defined in charset.h
+ * - confidence indicates the threshold for charset detection (0 to 1.0)
+ * - val holds any allocated memory to which the return value points to
+ * - (optional) is_encoding_problem is set for invalid byte sequences
+ *
+ * The return value MAY point to data if data is a C-string and does not
+ * contain invalid UTF-8 (but may contain replacement) characters.
+ */
+extern const char *jmap_decode_to_utf8(const char *charset, int encoding,
+                                       const char *data, size_t datalen,
+                                       float confidence,
+                                       char **val,
+                                       int *is_encoding_problem);
+
 #endif /* JMAP_UTIL_H */
